bob project: added project command to bob
With projectGenerators you can write generator plugins to create project files
for IDE's. The plugin entrys signature is:
    generator(package, args, extra)
Where package is the package to start from, args are arguments not consumed by bob
(they can be used inside the plugin), and extra are arguments you should give back
to bob when you call him from your IDE.

The plugin will get the root of the project as "bob.input.Package"
object. It also gets the rest of the command line and a list of
arguments that should be passed to "bob dev".

As a bob build-in generator the QtCreatorGenerator was added. This generator
generates a project with differen build settings. To have a single point of
information a buildme shell script is used to execute from QT-Creator.
It is is visible in the project tree and executes bob dev followed by bob project to update
project files. You can modify bob dev arguments as needed in this file.

All x86 executables in packageDir will be added as runTarget. bob project executes a bob dev
because binaries have to be build before generator can pick them up.